src/db.py
from __future__ import annotations

import os
from dataclasses import dataclass
from typing import Any, Iterable, Optional

import mysql.connector
from mysql.connector.connection import MySQLConnection

from dotenv import load_dotenv
load_dotenv()

@dataclass(frozen=True)
class DBConfig:
    """Configuración de conexión a la base de datos."""
    host: str
    port: int
    database: str
    user: str
    password: str


def load_config_from_env() -> DBConfig:
    """
    Lee la configuración de conexión desde variables de entorno.

    Variables esperadas (con valores por defecto):
    - DB_HOST (default: localhost)
    - DB_PORT (default: 3306)
    - DB_NAME (default: sti_incidencias)
    - DB_USER (default: sti_app)
    - DB_PASSWORD (default: sti_app_2026)

    Debe devolver un objeto DBConfig correctamente construido.

    Recomendación:
    - Validar que DB_PORT sea un número entero.
    """
    import os
from dataclasses import dataclass

@dataclass
class DBConfig:
    host: str
    port: int
    database: str
    user: str
    password: str


def load_config_from_env() -> DBConfig:
    """
    Lee la configuración de conexión desde variables de entorno.
    """

    host = os.getenv("DB_HOST", "localhost")
    port_str = os.getenv("DB_PORT", "3306")
    database = os.getenv("DB_NAME", "sti_incidencias")
    user = os.getenv("DB_USER", "sti_app")
    password = os.getenv("DB_PASSWORD", "sti_app_2026")

    # Validar que el puerto sea entero
    try:
        port = int(port_str)
    except ValueError:
        raise ValueError(f"DB_PORT debe ser un número entero, recibido: {port_str}")

    host= os.getenv("DB_HOST", "localhost")
    port= os.getenv("DB_PORT", "3306")
    database= os.getenv("DB_NAME", "sti_incidencias")
    user= os.getenv("DB_USER", "sti_app")
    password= os.getenv("DB_PASSWORD", "sti_app_2026")
    return DBConfig(host, port, database, user, password)

    raise NotImplementedError


def get_connection(cfg: Optional[DBConfig] = None) -> MySQLConnection:
    """
    Crea y devuelve una conexión MySQL/MariaDB.

    - Si cfg es None, debe llamar a load_config_from_env().
    - Debe usar mysql.connector.connect(...) con los parámetros de cfg.
    """
    import mysql.connector
from mysql.connector import MySQLConnection
from typing import Optional


def get_connection(cfg: Optional[DBConfig] = None) -> MySQLConnection:
    """
    Crea y devuelve una conexión MySQL/MariaDB.

    - Si cfg es None, llama a load_config_from_env().
    - Usa mysql.connector.connect(...) con los parámetros de cfg.
    """

    if cfg is None:
        cfg = load_config_from_env()

    conn = mysql.connector.connect(
        host=cfg.host,
        port=cfg.port,
        database=cfg.database,
        user=cfg.user,
        password=cfg.password,
    )

    return conn


    raise NotImplementedError


def fetch_all(conn: MySQLConnection, query: str, params: Optional[Iterable[Any]] = None) -> list[dict]:
    """
    Ejecuta una consulta SELECT y devuelve una lista de diccionarios (una fila -> un dict).

    Pistas:
    - Crear un cursor con conn.cursor(dictionary=True)
    - Ejecutar cur.execute(query, params o ())
    - Obtener filas con cur.fetchall()
    - Cerrar el cursor siempre (try/finally)
    """
    from typing import Optional, Iterable, Any
from mysql.connector import MySQLConnection


def fetch_all(
    conn: MySQLConnection,
    query: str,
    params: Optional[Iterable[Any]] = None
) -> list[dict]:
    """
    Ejecuta una consulta SELECT y devuelve una lista de diccionarios (una fila -> un dict).
    """

    cur = conn.cursor(dictionary=True)
    try:
        cur.execute(query, params or ())
        rows = cur.fetchall()
        return rows
    finally:
        cur.close()

    raise NotImplementedError


def execute(conn: MySQLConnection, query: str, params: Optional[Iterable[Any]] = None) -> int:
    """
    Ejecuta una sentencia INSERT/UPDATE/DELETE y devuelve el número de filas afectadas.

    Pistas:
    - Crear un cursor normal conn.cursor()
    - Ejecutar cur.execute(query, params o ())
    - Hacer conn.commit()
    - Devolver cur.rowcount
    - Cerrar el cursor siempre (try/finally)
    """
    from typing import Optional, Iterable, Any
from mysql.connector import MySQLConnection


def execute(
    conn: MySQLConnection,
    query: str,
    params: Optional[Iterable[Any]] = None
) -> int:
    """
    Ejecuta una sentencia INSERT/UPDATE/DELETE y devuelve el número de filas afectadas.
    """

    cur = conn.cursor()
    try:
        cur.execute(query, params or ())
        conn.commit()
        return cur.rowcount
    finally:
        cur.close()

    raise NotImplementedError
