requirements.txt
mysql-connector-python
pytest-dotenv
pytest
